fix: persist cached editor state synchronously
The cache write in WithCachedState ran inside the setState updater, so a
fast tab switch could unmount the editor before React flushed and drop the
write. The stale cache then triggered a re-import that discarded unsaved
changes.

Closes https://github.com/camunda/camunda-modeler/issues/6063
Closes https://github.com/camunda/camunda-modeler/issues/6064
